// components/common/product/type.ts
export interface IProductsProps {
   id: number;
   rate: number;
   slug: string;
   price: number;
   title: string;
   imgs: string[];
   discount: number;
   saleCount: number;
   createdAt: string;
   count: number;
   colors: string[];
   size: string[];
   description: string;
   category: string;
}